+++ concordium-node/src/grpc2.rs
@@ -987,8 +987,10 @@ pub mod server {
                         .context("Unable to start the GRPC2 reflection service.")?;
 
                     let health_service = health::HealthServiceImpl {
-                        consensus:                     consensus.clone(),
+                        consensus: consensus.clone(),
+                        node: node.clone(),
                         health_max_finalization_delay: config.health_max_finalized_delay,
+                        health_min_peers: config.health_min_peers,
                     };
                     router
                         .add_service(health::health_server::HealthServer::new(health_service))