+++ concordium-node/src/health.rs
@@ -1,7 +1,12 @@
 //! Implementation of the health check service that is part of the GRPC2
 //! interface.
 
-use crate::consensus_ffi::consensus::ConsensusContainer;
+use crate::{
+    common::PeerType,
+    consensus_ffi::{consensus::ConsensusContainer, helpers::ConsensusIsInBakingCommitteeResponse},
+    p2p::P2PNode,
+};
+use std::sync::Arc;
 
 include!(concat!(env!("OUT_DIR"), "/concordium.health.rs"));
 
@@ -10,8 +15,10 @@ pub(crate) static HEALTH_DESCRIPTOR: &[u8] =
 
 /// The type that implements the service that responds to queries.
 pub(crate) struct HealthServiceImpl {
-    pub(crate) consensus:                     ConsensusContainer,
+    pub(crate) consensus: ConsensusContainer,
+    pub(crate) node: Arc<P2PNode>,
     pub(crate) health_max_finalization_delay: concordium_base::base::DurationSeconds,
+    pub(crate) health_min_peers: Option<usize>,
 }
 
 #[tonic::async_trait]
@@ -40,6 +47,26 @@ impl health_server::Health for HealthServiceImpl {
 
         if delta > 1000 * u64::from(self.health_max_finalization_delay) {
             return Err(tonic::Status::unavailable("Last finalized block is too far behind."));
+        }
+
+        if let Some(min_allowed_peers) = self.health_min_peers {
+            let num_peers = self.node.get_peer_stats(Some(PeerType::Node)).len();
+            if num_peers < min_allowed_peers {
+                return Err(tonic::Status::unavailable(format!(
+                    "The only has {} peers, but is required to have at least {}.",
+                    num_peers, min_allowed_peers
+                )));
+            }
+        }
+
+        if self.consensus.is_active() {
+            let (committee_status, _, _, _) = self.consensus.in_baking_committee();
+            if committee_status != ConsensusIsInBakingCommitteeResponse::ActiveInCommittee {
+                return Err(tonic::Status::unavailable(
+                    "The node is configured with baker credentials, but is not in the baking \
+                     committee.",
+                ));
+            }
         }
 
         Ok(tonic::Response::new(NodeHealthResponse {}))