+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased changes
 
+- Add an option `--grpc2-health-min-peers` (environment variable `CONCORDIUM_NODE_GRPC2_HEALTH_MIN_PEERS`)
+  that causes the grpc V2 health endpoint to check minimum number of peers.
+- Extend the node health check so that if the node is configured with baker
+  credentials then it is required to be in the baking committee for it to be
+  considered healthy.
+
 ## 5.3.0
 
 - Extend Prometheus exporter with metrics: `grpc_request_duration_seconds`, `grpc_in_flight_requests`, `consensus_baking_committee`, `consensus_finalization_committee`, `consensus_baking_lottery_power`, `consensus_baked_blocks_total`, `consensus_finalized_baked_blocks_total`, `network_soft_banned_peers_total`, `consensus_non_finalized_transactions` and `consensus_unsupported_pending_protocol_version` see [docs/prometheus-exporter.md](https://github.com/Concordium/concordium-node/blob/main/docs/prometheus-exporter.md) for more details.